[refactor🔨] : request 수정
