+++ src/main/java/shop/gaship/gashipshoppingmall/category/controller/CategoryController.java
@@ -7,8 +7,8 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import shop.gaship.gashipshoppingmall.category.dto.CategoryDto;
-import shop.gaship.gashipshoppingmall.category.request.CategoryCreateRequest;
-import shop.gaship.gashipshoppingmall.category.request.CategoryModifyRequest;
+import shop.gaship.gashipshoppingmall.category.dto.CategoryCreateRequestDto;
+import shop.gaship.gashipshoppingmall.category.dto.CategoryModifyRequestDto;
 import shop.gaship.gashipshoppingmall.category.service.CategoryService;
 
 /**
@@ -37,7 +37,7 @@ public class CategoryController {
      * @return response entity
      */
     @PostMapping
-    public ResponseEntity<Void> postCategory(@RequestBody CategoryCreateRequest request) {
+    public ResponseEntity<Void> postCategory(@RequestBody CategoryCreateRequestDto request) {
         categoryService.createCategory(request);
 
         return ResponseEntity.status(HttpStatus.CREATED)
@@ -56,7 +56,7 @@ public class CategoryController {
      */
     @PutMapping("/{categoryNo}")
     public ResponseEntity<Void> putCategory(@PathVariable("categoryNo") Integer categoryNo,
-                                            @RequestBody CategoryModifyRequest request) {
+                                            @RequestBody CategoryModifyRequestDto request) {
         categoryService.modifyCategory(categoryNo, request);
 
         return ResponseEntity.status(HttpStatus.NO_CONTENT)
